Adds a deep clean test with and without beam
This test only checks the peak dirty to peak final residue against
the corresponding value of the reference dataset and skips
maximum squared and mean squared error checks becuase we will likely
clean very deeply into the sidelobes which may change with small changes
in the code base.

There is only one source in this field so we can clean relatively quickly.
The source also has a very steep spectral slope, so coupled with the fact
that it is at the edge of the beam should make for an interresting test.